Refactor parsing logic into knowpro
