+++ ts/packages/knowPro/test/testCommon.ts
@@ -29,7 +29,7 @@ import {
 import { createConversationFromData } from "../src/common.js";
 import { readConversationDataFromFile } from "../src/serialization.js";
 import { SemanticRefSearchResult } from "../src/search.js";
-import { createSearchTerm } from "../src/searchCommon.js";
+import { createSearchTerm } from "../src/searchLib.js";
 import * as q from "../src/query.js";
 import { PropertyNames } from "../src/propertyIndex.js";
 import { Result } from "typechat";
